requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
psycopg2-binary
mysqlclient
Pillow
python-dotenv
gunicorn
werkzeug<3.1
click
blinker
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
WTForms
email-validator 